# Benchmark some sampling cases